+++ CHANGELOG.md
@@ -6,7 +6,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 ### New
 
-- Lint arities of fn arguments to `map`, `filter`, `reduce` and related  [#1297](https://github.com/clj-kondo/clj-kondo/issues/1297)
+- Lint arities of fn arguments to `map`, `filter`, `reduce` and related [#1297](https://github.com/clj-kondo/clj-kondo/issues/1297)
 
 ### Enhanced / fixed